feat(Button): convert Button to typescript
